requirements.txt
feedparser==6.0.10
requests==2.32.4
python-dotenv==1.0.1
pillow==10.1.0
supabase==2.4.6
cryptography==46.0.3
groq==0.4.2
# Instagram Graph API - using requests library (no extra dependencies needed)

# New dependencies for 2-3 posts/hour system
translatepy==2.3
apscheduler==3.10.4
pytz==2024.1
fastapi==0.109.0
uvicorn[standard]==0.27.0